+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) 
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## 2.0.0
+
+* [CHANGED] - requires bi-service >= 1.0.0
+* [ADDED] - `spread` query parameter to the GET `api/v1.0/integrity` route
+
 ## 2.0.0-alpha
 
 * [CHANGED] - requires bi-service >= 1.0.0-alpha